acceleration ramp-based path planner
This commit replaces most of the guts of the path planner. It replaces
Bresenham's line algorithm with a system based on ideal between-step delays
and "acceleration ramps," as explained in the code.

This change has the immediate effect that the path planner is more stable at
extremes and generates cleaner speed trapezoids. For example, the old path
planner had a worst case where one axis was moving just a bit slower than
the primary axis for a move, in which case the slower axis would stutter
as it alternated between stepping a bit too quickly and missing steps.
The old path planner also used a fixed point approximation for acceleration
such that if acceleration was too low, an internal value would go to zero
and the move would jump from start speed to full speed to end speed with
no acceleration at all.

This change may have a minor effect on print quality, but it isn't itself
the end goal. Calculating axes independently and based on ideal step delays
makes other things easier:
- pressure advance - The extruder path can now be modified independently
    of the other axes. I've prototyped a few ways of implementing pressure
    advance, and the only difficult part left is making it obey jerk and maximum
    acceleration (which other firmwares generally don't do).
- delta moves - Bresenham assumes the ratio of movement between multiple
   axes is constant for the duration of a move. With this assumption removed,
   it is no longer necessary for Delta moves to be broken into tiny pieces.
   This will bring improved print quality because moves can accelerate smoothly
   rather than relying in jerk as they do now.

+++ redeem/path_planner/test_harness/MockPruTimer.cpp
@@ -79,7 +79,7 @@ void PruTimer::push_block(uint8_t* blockMemory, size_t blockLen, unsigned int un
   renderedPath.path = path;
 
   renderedPath.stepperCommands.reserve(blockLen / sizeof(SteppersCommand));
-  for (int i = 0; i < blockLen / sizeof(SteppersCommand); i++) {
+  for (size_t i = 0; i < blockLen / sizeof(SteppersCommand); i++) {
     renderedPath.stepperCommands.push_back(commands[i]);
   }
